Move the LogicLayer class to its own module.
